many improvements * added height to NN (pi-sigma networks) * input_vector for ranking nn now created using standard methods * evolution of embedding for current track * experiments on which action to perform in ranker (run cma-es inside, or improve previous action) * add method to guaranteedly evolve working NN (evolve_many_cmaes) * experiments on random NNs * experiments on NN height

+++ src/math.rs
@@ -9,10 +9,10 @@ use std::ops::Neg;
 use std::ops::Sub;
 use std::ops::SubAssign;
 
-pub type fxx = f64;
-pub const PIx: fxx = std::f64::consts::PI;
-pub const TAUx: fxx = std::f64::consts::TAU;
-pub const Ex: fxx = std::f64::consts::E;
+pub type fxx = f32;
+pub const PIx: fxx = std::f32::consts::PI;
+pub const TAUx: fxx = std::f32::consts::TAU;
+pub const Ex: fxx = std::f32::consts::E;
 
 #[derive(Clone, Copy, Default, PartialEq, serde::Deserialize, serde::Serialize)]
 pub struct Vecx2 {
